requirements.txt
streamlit==1.37.1
pandas
numpy
matplotlib
python-dotenv
bcrypt
cffi
passlib[bcrypt]
openai